requirements.txt
Pillow
pytest
qrcode